UI Conversation

+++ vyger-app/dist.ps1
@@ -4,6 +4,10 @@ $root = "$PSScriptRoot"
 $dist = "$root\dist\vyger-app"
 $docs = [System.IO.Path]::GetFullPath("$root\..\docs")
 
+if (-not(Test-Path -Path $dist)) {
+    New-Item -Path $dist -ItemType Directory | Out-Null
+}
+
 Write-Host $docs
 
 if (Test-Path -Path $docs) {
@@ -14,7 +18,12 @@ else {
 }
 
 Write-Host "Building .. " -ForegroundColor Green
-&ng build --prod --base-href /vyger/
+&ng build --prod --base-href ./
 
-Write-Host "Copying to docs .. " -ForegroundColor Yellow
-Copy-Item -Path "$dist\*.*" -Destination "$docs\"
+if (Test-Path -Path $dist) {
+    Write-Host "Copying to docs .. " -ForegroundColor Yellow
+    Copy-Item -Path "$dist\*.*" -Destination "$docs\"
+}
+else {
+    Write-Error "Missing $dist"
+}
